backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
groq>=0.4.0
httpx>=0.25.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
sqlalchemy[asyncio]>=2.0.23
aiosqlite>=0.19.0
alembic>=1.13.0
